Gene synonyms

Look up gene synonyms from NCBI.

geneSynonyms(
  organism = c("Homo sapiens", "Mus musculus", "Drosophila melanogaster"),
  return = c("DataFrame", "SplitDataFrameList")
)

Arguments

organism
character(1). Full Latin organism name (e.g. "Homo sapiens").

return
character(1). Return type. Uses match.arg() internally and defaults to the first argument in the character vector.

Value

DataFrame: Returns unique row for each geneID. Returns only geneID and geneSynonyms columns. The geneSynonyms column returns as character vector, with synonyms arranged alphabetically and delimited by ", ".

SplitDataFrameList: Split by geneID column. Keeps duplicate rows mapping to same geneID (e.g. ENSG00000004866). Returns geneID, geneName, and geneSynonyms columns in the split.

Note

Synonym support for Caenorhabditis elegans is poor on NCBI. Use the wormbase package instead.

Updated 2020-10-05.

Examples

object <- geneSynonyms(organism = "Homo sapiens")

#> → Importing `Homo sapiens` gene synonyms from Ensembl.

#> → Caching URL at <ftp://ftp.ncbi.nih.gov/gene/DATA/GENE_INFO/Mammalia/Homo_sapiens.gene_info.gz>.

#> → Importing 13da01b767cf8_Homo_sapiens.gene_info.gz at /Users/mike/Library/Caches/AcidGenomes using vroom::`vroom()`.

#> → Preparing unique synonyms per gene.

print(object)

#> DataFrame with 22715 rows and 2 columns
#>                          geneID
#>                     <character>
#> ENSG00000000003 ENSG00000000003
#> ENSG00000000005 ENSG00000000005
#> ENSG00000000419 ENSG00000000419
#> ENSG00000000457 ENSG00000000457
#> ENSG00000000938 ENSG00000000938
#> ...                         ...
#> ENSG00000288601 ENSG00000288601
#> ENSG00000288611 ENSG00000288611
#> ENSG00000288621 ENSG00000288621
#> ENSG00000288638 ENSG00000288638
#> ENSG00000288642 ENSG00000288642
#>                                                                   geneSynonyms
#>                                                                    <character>
#> ENSG00000000003                                  T245, TM4SF6, TSPAN-6, TSPAN6
#> ENSG00000000005                                       BRICD4, CHM1L, TEM, TNMD
#> ENSG00000000419                                              CDGIE, DPM1, MPDS
#> ENSG00000000457                                           PACE-1, PACE1, SCYL3
#> ENSG00000000938 c-fgr, c-src2, FGR, p55-Fgr, p55c-fgr, p58-Fgr, p58c-fgr, SRC2
#> ...                                                                        ...
#> ENSG00000288601                                             RN5S424, RNA5SP424
#> ENSG00000288611                                                   GPR7, NPBWR1
#> ENSG00000288621                                          XG, XGPY, XGPY2, XGY2
#> ENSG00000288638                                                 LNCDAT, YYLNCT
#> ENSG00000288642                                       CDR, CDR1, CDR34, CDR62A